Obliterate @SubcircuitPin from existence (and links to my repo)
Probably due to a botched rebase (which I arguably inflicted on the TAs,
my apologies), @SubcircuitPin is incorrectly still present in this
repository despite CircuitSimExtension not caring about it. Remove it
from this repository and the readme. Also remove links to my repository
because it's behind and might confuse people.